modified:   CMP_N7/src/modules/FormEditPlayer.jsx

+++ CMP_N7/src/modules/FormEditPlayer.jsx
@@ -22,10 +22,10 @@ function FormEditPlayer({ onClose, Data, UpdateData }) {
   function onNameChange (newText) { setNombreJugador(newText) }
   function onCLT (newText) { setLT(parseInt(newText)) }
   function onCLR (newText) { setLR(parseInt(newText)) }
-  function onCLB (newText) { setLB(parseInt(newText)) }
+  function onCLB (newText) { setLB(parseFloat(newText)) }
   function onCST (newText) { setST(parseInt(newText)) }
   function onCSR (newText) { setSR(parseInt(newText)) }
-  function onCSB (newText) { setSB(parseInt(newText)) }
+  function onCSB (newText) { setSB(parseFloat(newText)) }
   function onOK () {
     const newData = {...Data}
     newData.playerName = nombreJugador
